Add change algorithm, without None
